Change hook name for Webpack 4
Using 'compilation' I was receiving many 'Building X' logs before a single build completed. I tried this on a few projects and 'compile' seems to instead be correct. `compile` is also what is used on Webpack <= 3.x and as far as I can tell nothing changed with that event between 3 and 4 so they should probably be consistent anyways.